# backend/docs/openapi.yaml
openapi: 3.0.0
info:
  title: Foro-T API
  version: 1.0.0
  description: API para foro de traducción de videojuegos
servers:
  - url: http://localhost:5000/api
    description: Servidor de desarrollo
paths:
...
  /games:
    get:
      tags:
        - Games
      summary: Obtener lista de juegos
      responses:
        200:
          description: Lista de juegos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Game'
    
    post:
      tags:
        - Games
      summary: Crear nuevo juego
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                platform:
                  type: array
                  items:
                    type: string
      responses:
        201:
          description: Juego creado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Game'
...
  /api/forum/threads:
    post:
      tags:
        - Forum
      summary: Crear nuevo hilo
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Thread'
      responses:
        201:
          description: Hilo creado exitosamente
        403:
          description: Karma insuficiente
    
    get:
      tags:
        - Forum
      summary: Obtener todos los hilos
      responses:
        200:
          description: Lista de hilos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Thread'
  /api/forum/posts:
    post:
      tags:
        - Forum
      summary: Crear post en hilo
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ForumPost'
      responses:
        201:
          description: Post creado exitosamente
        403:
          description: Karma insuficiente
  /api/forum/comments:
    post:
      tags:
        - Forum
      summary: Crear comentario
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ForumComment'
      responses:
        201:
          description: Comentario creado exitosamente
        403:
          description: Karma insuficiente para comentar
...
  /api/forum/posts/{id}/vote:
    post:
      tags:
        - Forum
      summary: Votar un post
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                type:
                  type: string
                  enum: [up, down]
      responses:
        200:
          description: Voto registrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ForumPost'
        403:
          description: Karma insuficiente para votar
  /api/forum/comments/{postId}:
    get:
      tags:
        - Forum
      summary: Obtener comentarios de un post
      parameters:
        - in: path
          name: postId
          required: true
          schema:
            type: string
      responses:
        200:
          description: Lista de comentarios
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ForumComment'
  /api/forum/comments/{id}/vote:
    post:
      tags:
        - Forum
      summary: Votar un comentario
      description: Permite votar positiva o negativamente un comentario
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
          description: ID del comentario
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                type:
                  type: string
                  enum: [up, down]
                  description: Tipo de voto
      responses:
        200:
          description: Voto registrado exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ForumComment'
        403:
          description: Karma insuficiente para votar
        404:
          description: Comentario no encontrado
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    User:
      type: object
      required:
        - username
        - email
      properties:
        username:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [user, translator, moderator, admin]
        reputation:
          type: number
        level:
          type: string
    Game:
      type: object
      required:
        - title
        - platform
      properties:
        _id:
          type: string
        title:
          type: string
        platform:
          type: array
          items:
            type: string
            enum: [PC, PS4, PS5, Xbox One, Xbox Series, Switch, Mobile]
        status:
          type: string
          enum: [pending, in_progress, completed]
        translationProgress:
          type: number
          minimum: 0
          maximum: 100
        translationLead:
          $ref: '#/components/schemas/User'
        translators:
          type: array
          items:
            type: object
            properties:
              user:
                $ref: '#/components/schemas/User'
              assignedDate:
                type: string
                format: date-time
...
    Thread:
      type: object
      required:
        - name
        - description
      properties:
        name:
          type: string
          description: Nombre del hilo
        description:
          type: string
          description: Descripción del hilo
        type:
          type: string
          enum: [official, community]
        creator:
          $ref: '#/components/schemas/User'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    ForumPost:
      type: object
      required:
        - title
        - content
      properties:
        title:
          type: string
        content:
          type: string
        author:
          $ref: '#/components/schemas/User'
        thread:
          $ref: '#/components/schemas/Thread'
        votes:
          type: object
          properties:
            up:
              type: array
              items:
                $ref: '#/components/schemas/User'
            down:
              type: array
              items:
                $ref: '#/components/schemas/User'
    ForumComment:
      type: object
      required:
        - content
      properties:
        content:
          type: string
        author:
          $ref: '#/components/schemas/User'
        post:
          $ref: '#/components/schemas/ForumPost'
        votes:
          type: object
          properties:
            up:
              type: array
              items:
                $ref: '#/components/schemas/User'
            down:
              type: array
              items:
                $ref: '#/components/schemas/User'
...