# backend/forum/models/ForumComment.js
/**
 * @swagger
 * components:
 *   schemas:
 *     ForumComment:
 *       type: object
 *       required:
 *         - content
 *         - author
 *         - post
 *       properties:
 *         content:
 *           type: string
 *           description: Contenido del comentario
 *         author:
 *           $ref: '#/components/schemas/User'
 *         post:
 *           $ref: '#/components/schemas/ForumPost'
 *         parentComment:
 *           $ref: '#/components/schemas/ForumComment'
 *         votes:
 *           type: object
 *           properties:
 *             up:
 *               type: array
 *               items:
 *                 $ref: '#/components/schemas/User'
 *             down:
 *               type: array
 *               items:
 *                 $ref: '#/components/schemas/User'
 */
import mongoose from 'mongoose';

const forumCommentSchema = new mongoose.Schema({
    content: {
        type: String,
        required: true
    },
    author: {
        type: mongoose.Schema.Types.ObjectId,
        ref: 'User'
    },
    post: {
        type: mongoose.Schema.Types.ObjectId,
        ref: 'ForumPost'
    },
    parentComment: {
        type: mongoose.Schema.Types.ObjectId,
        ref: 'ForumComment'
    },
    votes: {
        up: [{
            type: mongoose.Schema.Types.ObjectId,
            ref: 'User'
        }],
        down: [{
            type: mongoose.Schema.Types.ObjectId,
            ref: 'User'
        }]
    }
}, { 
    timestamps: true,
    toJSON: { virtuals: true }
});

forumCommentSchema.virtual('replies', {
    ref: 'ForumComment',
    localField: '_id',
    foreignField: 'parentComment'
});

const ForumComment = mongoose.model('ForumComment', forumCommentSchema);
export default ForumComment;